order update

+++ tai-shop/Models/Item.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         //Orders
         public List<ItemOrder> ItemOrders { get; set; }
         //Return
@@ -14,6 +14,7 @@
         public List<ItemTag> ItemTags { get; set; }
         //Photos
         public List<Photo> Photos { get; set; } = new List<Photo>();
+        //Reviews
         public List<Review> Reviews { get; set; }
         public double AverageRating => Reviews?.Any() == true
             ? Reviews.Average(r => r.Rating)
